+++ internal/provider/cloudconnect_supported_features_data_source_test.go
@@ -1,0 +1,20 @@
+package provider_test
+
+import (
+	"regexp"
+	"testing"
+
+	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
+)
+
+func TestAccCloudconnectSupportedFeaturesDataSource_NotFound(t *testing.T) {
+	resource.ParallelTest(t, resource.TestCase{
+		ProtoV6ProviderFactories: testAccProvidersProtoV6Factories,
+		PreCheck:                 testAccPreCheckFunc(t),
+		TerraformVersionChecks:   testAccTFVersionChecks,
+		Steps: []resource.TestStep{{
+			Config:      `data "doit_cloudconnect_supported_features" "missing" { account_id = "000000000000" }`,
+			ExpectError: regexp.MustCompile(`status: 404`),
+		}},
+	})
+}